backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
supabase==2.3.5
langchain==0.2.16
langchain-community==0.2.17
langchain-openai==0.1.25
tavily-python==0.3.3
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
